pyproject.toml
[project]
name = "note-bot"
version = "0.1.0"
description = "Discord Bot for automated LaTeX note-taking via LLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.4.0",
    "PyGithub>=2.3.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]
